+++ src/board_representation/move_gen.rs
@@ -1,0 +1,156 @@
+use crate::board_representation::precompiled_bitboards::*;
+
+
+// from white's perspective
+const TOP_RANK: u64 = 0xFF00000000000000;
+const BOTTOM_RANK: u64 = 0x00000000000000FF;
+const LEFT_FILE: u64 = 0x8080808080808080;
+const RIGHT_FILE: u64 = 0x0101010101010101;
+
+const RANK_SHIFT: i32 = 8; // value to shift if you want to move ranks
+const FILE_SHIFT: i32 = 1; // value to shift if you want to move files
+
+pub fn pawn_moves(bitboard: &u64, friendly_bitboard: &u64, enemy_bitboard: &u64, is_white: bool)  -> u64 { // TODO: promotion, en_passent
+  let all_pieces = friendly_bitboard | enemy_bitboard;
+  let mut moves: u64 = 0;
+  let mut attacks: u64 = 0;
+
+  if is_white {
+    moves |= bitboard << RANK_SHIFT;
+    if bitboard & (BOTTOM_RANK << RANK_SHIFT) != 0 { // if pawn is on 2nd rank
+      moves |= bitboard << (RANK_SHIFT * 2);
+    }
+
+    if bitboard & RIGHT_FILE == 0 { // if piece is not on the left file
+      attacks |= bitboard << (RANK_SHIFT - 1)
+    }
+    if bitboard & LEFT_FILE == 0 { // if piece is not on the right file
+      attacks |= bitboard << (RANK_SHIFT + 1);
+    }
+  } else {
+    moves |= bitboard >> RANK_SHIFT;
+    if bitboard & (TOP_RANK >> RANK_SHIFT) != 0 { // if pawn is on 7th rank
+      moves |= bitboard >> (RANK_SHIFT * 2);
+    }
+
+    if bitboard & RIGHT_FILE == 0 { // if piece is not on the left file
+      attacks |= bitboard >> (RANK_SHIFT + 1)
+    }
+    if bitboard & LEFT_FILE == 0 { // if piece is not on the right file
+      attacks |= bitboard >> (RANK_SHIFT - 1);
+    }
+  }
+
+  moves ^= all_pieces & moves; // removes squares where another piece is. doesnt affect the pawn attacks
+  attacks ^= attacks & friendly_bitboard; // removes attacks on friendly pieces
+  if attacks & all_pieces == 0 { // if the pawn attacks nothing
+    attacks = 0; // attacks mean nothing
+  }
+
+  moves |= attacks;
+  moves
+}
+
+pub fn knight_moves(bitboard: &u64, friendly_bitboard: &u64) -> u64 {
+  let mut moves = 0;
+
+  if (bitboard & TOP_RANK == 0) && (bitboard & (LEFT_FILE | (LEFT_FILE >> FILE_SHIFT)) == 0) { // if not on top rank AND if not on the two left-most files\
+    moves |= bitboard << 10; // up left left
+  }
+  if (bitboard & (TOP_RANK & (TOP_RANK >> RANK_SHIFT)) == 0) && (bitboard & LEFT_FILE == 0) { // if not on the two top-most ranks AND if not on the left file
+    moves |= bitboard << 17; // up up left
+  }
+  if (bitboard & (TOP_RANK & (TOP_RANK >> RANK_SHIFT)) == 0) && (bitboard & RIGHT_FILE == 0) { // if not on the two top-most ranks AND if not on the right file
+    moves |= bitboard << 15; // up up right
+  }
+  if (bitboard & TOP_RANK == 0) && (bitboard & (RIGHT_FILE | (RIGHT_FILE << FILE_SHIFT)) == 0) { // if not on top rank AND if not on the two right-most files
+    moves |= bitboard << 6; // up right right
+  }
+  if (bitboard & BOTTOM_RANK == 0) && (bitboard & (RIGHT_FILE | (RIGHT_FILE << FILE_SHIFT)) == 0) { // if not on bottom rank AND if not on the two right-most files
+    moves |= bitboard >> 10; // down right right
+  }
+  if (bitboard & (BOTTOM_RANK & (BOTTOM_RANK << RANK_SHIFT)) == 0) && (bitboard & RIGHT_FILE == 0) { // if not on the two bottom-most ranks AND if not on the right file
+    moves |= bitboard >> 17; // down down right
+  }
+  if (bitboard & (BOTTOM_RANK & (BOTTOM_RANK << RANK_SHIFT)) == 0) && (bitboard & LEFT_FILE == 0) { // if not on the two bottom-most ranks AND if not on the left file
+    moves |= bitboard >> 15; // down down left
+  }
+  if (bitboard & BOTTOM_RANK == 0) && (bitboard & (LEFT_FILE | (LEFT_FILE >> FILE_SHIFT)) == 0) { // if not on bottom rank AND if not on the two left-most files
+    moves |= bitboard >> 6; // down left left
+  }
+  
+  let false_attacks = moves & friendly_bitboard;
+  if false_attacks != 0 {
+    moves ^= false_attacks;
+  }
+
+  moves
+}
+pub fn king_moves(bitboard: &u64, friendly_bitboard: &u64) -> u64 { // TODO: castling
+  let mut moves = 0;
+
+  if bitboard & TOP_RANK == 0 { // if not on the top of the board
+    moves |= bitboard << RANK_SHIFT; // up
+    
+    if bitboard & RIGHT_FILE == 0 { // if not on the right of the board
+      moves |= bitboard << RANK_SHIFT - 1; // up right
+    }
+    if bitboard & LEFT_FILE == 0 { // if not on the left of the board
+      moves |= bitboard << RANK_SHIFT + 1; // up left
+    }
+  }
+  if bitboard & BOTTOM_RANK == 0 { // if not on the bottom of the board
+    moves |= bitboard >> RANK_SHIFT; // down
+
+    if bitboard & LEFT_FILE == 0 { // if not on the left of the board
+      moves |= bitboard >> RANK_SHIFT - 1; // down left
+    }
+    if bitboard & RIGHT_FILE == 0 { // if not on the right of the board
+      moves |= bitboard >> RANK_SHIFT + 1; // down right
+    }
+  }
+  if bitboard & LEFT_FILE == 0 { // if not on the left of the board
+    moves |= bitboard << FILE_SHIFT; // left
+  }
+  if bitboard & RIGHT_FILE == 0 { // if not on the right of the board
+    moves |= bitboard >> FILE_SHIFT; // right
+  }
+  
+  let false_attacks = moves & friendly_bitboard;
+  if false_attacks != 0 {
+    moves ^= false_attacks;
+  }
+
+  moves
+}
+
+pub fn get_magic_index(magic: u64, index_bits: u32, mask: u64, population: &u64) -> usize {
+  let blockers = population & mask;
+
+  (blockers.wrapping_mul(magic) >> index_bits) as usize
+}
+
+pub fn get_bishop_moves(square_index: i32, friendly_bitboard: &u64, enemy_bitboard: &u64) -> u64 {
+  let population = friendly_bitboard | enemy_bitboard;
+  
+  let magic = &BISHOP_MAGICS[square_index as usize];
+  let mask = &BISHOP_MASKS[square_index as usize];
+  let relevant_bits = &BISHOP_BITS[square_index as usize];
+
+  let mut moves = BISHOP_MOVES[square_index as usize][get_magic_index(*magic, *relevant_bits, *mask, &population)];
+  moves ^= friendly_bitboard & mask;
+  
+  moves
+}
+pub fn get_rook_moves(square_index: i32, friendly_bitboard: &u64, enemy_bitboard: &u64) -> u64 {
+  let population = friendly_bitboard | enemy_bitboard;
+  
+  let magic = &ROOK_MAGICS[square_index as usize];
+  let mask = &ROOK_MASKS[square_index as usize];
+  let relevant_bits = &ROOK_BITS[square_index as usize];
+
+  let mut moves = ROOK_MOVES[square_index as usize][get_magic_index(*magic, *relevant_bits, *mask, &population)];
+  moves ^= friendly_bitboard & mask;
+  
+  moves
+}
